src/requirements.txt
accelerate~=0.21.0
bitsandbytes~=0.40.2
evaluate~=0.4.0
hydra-core~=1.3.2
nltk~=3.8.1
numpy~=1.25.1
pandas~=2.0.3
peft~=0.4.0
scipy~=1.11.1
torch~=2.0.1
transformers~=4.31.0
xformers~=0.0.20

# Development dependencies
mypy~=1.4.1
autopep8~=2.0.2
isort~=5.12.0
jupyter~=1.0.0
